requirements.txt
flask
rasa
Flask-PyMongo
requests